requirements.txt
Flask==2.3.2
openai==1.50.2
python-dotenv==1.0.0
pandas==2.2.3
openpyxl==3.1.5